requirements.txt
lxml==3.4.4
python-magic==0.4.6
wheel==0.24.0
